services/service_c/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
pydantic[email]==2.6.1
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.1
PyJWT==2.8.0
bcrypt==4.1.2
jinja2==3.1.3
aiofiles==23.2.1
PyPDF2==3.0.1
numpy<2.0
openai==1.12.0
